i can explain.
i will not be removing revisions of this file from commit history, because that would symbolize "i refuse to admit this ever happened."

i'm not a perfect person. nobody is. we all make some very bad mistakes sometimes. i'm sorry what what i've done. (the people in question have forgiven me and we are on good terms now. life is good.)